+++ src/main/java/com/codecool/advanced_project/repository/CustomTagEntityRepository.java
@@ -3,9 +3,9 @@ package com.codecool.advanced_project.repository;
 import com.codecool.advanced_project.entity.CustomTagEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomTagEntityRepository extends JpaRepository<CustomTagEntity, Long> {
+public interface CustomTagEntityRepository extends JpaRepository<CustomTagEntity, Integer> {
 
-    CustomTagEntity findByIdAnd(Long id);
+    CustomTagEntity findByIdAnd(Integer id);
 
     CustomTagEntity findByName(String name);
 }
